vcmi: split selectors from HeroBonus.h
