+++ lib/bonuses/BonusList.h
@@ -10,6 +10,7 @@
 #pragma once
 
 #include "HeroBonus.h"
+#include "BonusSelector.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
