+++ lib/bonuses/HeroBonus.cpp
@@ -354,79 +354,6 @@ std::shared_ptr<Bonus> Bonus::addPropagator(const TPropagatorPtr & Propagator)
 	return this->shared_from_this();
 }
 
-namespace Selector
-{
-	DLL_LINKAGE CSelectFieldEqual<Bonus::BonusType> & type()
-	{
-		static CSelectFieldEqual<Bonus::BonusType> stype(&Bonus::type);
-		return stype;
-	}
-
-	DLL_LINKAGE CSelectFieldEqual<TBonusSubtype> & subtype()
-	{
-		static CSelectFieldEqual<TBonusSubtype> ssubtype(&Bonus::subtype);
-		return ssubtype;
-	}
-
-	DLL_LINKAGE CSelectFieldEqual<CAddInfo> & info()
-	{
-		static CSelectFieldEqual<CAddInfo> sinfo(&Bonus::additionalInfo);
-		return sinfo;
-	}
-
-	DLL_LINKAGE CSelectFieldEqual<Bonus::BonusSource> & sourceType()
-	{
-		static CSelectFieldEqual<Bonus::BonusSource> ssourceType(&Bonus::source);
-		return ssourceType;
-	}
-
-	DLL_LINKAGE CSelectFieldEqual<Bonus::BonusSource> & targetSourceType()
-	{
-		static CSelectFieldEqual<Bonus::BonusSource> ssourceType(&Bonus::targetSourceType);
-		return ssourceType;
-	}
-
-	DLL_LINKAGE CSelectFieldEqual<Bonus::LimitEffect> & effectRange()
-	{
-		static CSelectFieldEqual<Bonus::LimitEffect> seffectRange(&Bonus::effectRange);
-		return seffectRange;
-	}
-
-	DLL_LINKAGE CWillLastTurns turns;
-	DLL_LINKAGE CWillLastDays days;
-
-	CSelector DLL_LINKAGE typeSubtype(Bonus::BonusType Type, TBonusSubtype Subtype)
-	{
-		return type()(Type).And(subtype()(Subtype));
-	}
-
-	CSelector DLL_LINKAGE typeSubtypeInfo(Bonus::BonusType type, TBonusSubtype subtype, const CAddInfo & info)
-	{
-		return CSelectFieldEqual<Bonus::BonusType>(&Bonus::type)(type)
-			.And(CSelectFieldEqual<TBonusSubtype>(&Bonus::subtype)(subtype))
-			.And(CSelectFieldEqual<CAddInfo>(&Bonus::additionalInfo)(info));
-	}
-
-	CSelector DLL_LINKAGE source(Bonus::BonusSource source, ui32 sourceID)
-	{
-		return CSelectFieldEqual<Bonus::BonusSource>(&Bonus::source)(source)
-			.And(CSelectFieldEqual<ui32>(&Bonus::sid)(sourceID));
-	}
-
-	CSelector DLL_LINKAGE sourceTypeSel(Bonus::BonusSource source)
-	{
-		return CSelectFieldEqual<Bonus::BonusSource>(&Bonus::source)(source);
-	}
-
-	CSelector DLL_LINKAGE valueType(Bonus::ValueType valType)
-	{
-		return CSelectFieldEqual<Bonus::ValueType>(&Bonus::valType)(valType);
-	}
-
-	DLL_LINKAGE CSelector all([](const Bonus * b){return true;});
-	DLL_LINKAGE CSelector none([](const Bonus * b){return false;});
-}
-
 DLL_LINKAGE std::ostream & operator<<(std::ostream &out, const Bonus &bonus)
 {
 	for(const auto & i : bonusNameMap)
